solution2
