+++ src/App.tsx
@@ -1,13 +1,20 @@
 import React from 'react';
 import { Clock } from './Clock';
+import './App.scss';
 
 interface State {
   hasClock: boolean;
   clockName: string;
 }
 
+function getRandomName(): string {
+  const value = Date.now().toString().slice(-4);
+
+  return `Clock-${value}`;
+}
+
 export class App extends React.Component<{}, State> {
-  timerID: NodeJS.Timeout | null = null;
+  nameTimerID: NodeJS.Timeout | undefined = undefined;
 
   state: State = {
     hasClock: true,
@@ -17,11 +24,19 @@ export class App extends React.Component<{}, State> {
   componentDidMount() {
     document.addEventListener('contextmenu', this.handleContextMenu);
     document.addEventListener('click', this.handleMouseClick);
+    this.nameTimerID = setInterval(() => this.updateClockName(), 3300);
   }
 
   componentWillUnmount() {
     document.removeEventListener('contextmenu', this.handleContextMenu);
     document.removeEventListener('click', this.handleMouseClick);
+    clearInterval(this.nameTimerID);
+  }
+
+  updateClockName() {
+    const newName = getRandomName();
+
+    this.setState({ clockName: newName });
   }
 
   handleContextMenu = (event: MouseEvent) => {
